Updated changes on 29 dec 2023 by Shubham

+++ app/src/main/java/com/ab/hicareservices/data/model/servicesmodule/ValidateServiceVoucherResponseData.kt
@@ -3,7 +3,7 @@ package com.ab.hicareservices.data.model.servicesmodule
 import com.google.gson.annotations.SerializedName
 
 data class ValidateServiceVoucherResponseData(
-    @SerializedName("VoucherDiscount") var VoucherDiscount: Int? = null,
+    @SerializedName("VoucherDiscount") var VoucherDiscount: Double? = null,
     @SerializedName("VoucherDiscount_InPercentage") var VoucherDiscountInPercentage: Int? = null,
-    @SerializedName("FinalAmount") var FinalAmount: Int? = null
+    @SerializedName("FinalAmount") var FinalAmount: Double? = null
 )
